+++ controllers/listingstories.js
@@ -6,44 +6,40 @@ const { cleanseData, parseSqlUpdateStmt } = require('../utils/dbHelper');
 /**
  * @desc    Get all listing stories
  * @route   GET /api/listings/stories
+ * @desc    Get single listing story
+ * @route   GET /api/listings/:listing_id/stories
  * @access  Public
  */
 exports.getListingStories = asyncHandler(async (req, res) => {
+  if (req.params.listing_id) {
+    const rows = await db.one(
+      'SELECT * FROM listingstories WHERE listing_id = $1',
+      req.params.listing_id
+    );
+    return res.status(200).json({
+      success: true,
+      data: rows,
+    });
+  }
   res.status(200).json(res.advancedResults);
 });
 
 /**
- * @desc    Get single listing story
- * @route   GET /api/listings/stories/:id
- * @access  Public
- */
-exports.getListingStory = asyncHandler(async (req, res) => {
-  const rows = await db.one(
-    'SELECT * FROM listingstories WHERE listing_id = $1',
-    req.params.id
-  );
-  res.status(200).json({
-    success: true,
-    data: rows,
-  });
-});
-
-/**
  * @desc    Update single listing story
- * @route   PUT /api/listings/stories/:id
+ * @route   PUT /api/listings/:listing_id/stories
  * @access  Admin/Owner
  */
 exports.updateListingStory = asyncHandler(async (req, res, next) => {
   // check if listing exists
   const listing = await db.one(
     'SELECT * FROM listings WHERE listing_id = $1',
-    req.params.id
+    req.params.listing_id
   );
-  // if non-admin user, throw 403 if not updating self
+  // if non-admin user, throw 403 if not listing owner
   if (req.user.role !== 'admin' && req.user.user_id !== listing.created_by) {
     return next(
       new ErrorResponse(
-        `Not allowed to update story of listing which you did not create`,
+        `Not allowed to update story of listing which you are not the owner of`,
         403
       )
     );
@@ -64,7 +60,7 @@ exports.updateListingStory = asyncHandler(async (req, res, next) => {
     data,
     'listingstories',
     'WHERE listing_id = $1 RETURNING $2:name',
-    [req.params.id, data]
+    [req.params.listing_id, data]
   );
 
   const rows = await db.one(updateListingStoryQuery);